Add files via upload
This is a bash script that uses the nmap tool to scan a network and determine if a host is up or not. It prompts the user to input a target IP address or range, then uses nmap with the "-sP" option to perform a fast ping scan of the target network. The results of the scan are saved to a file named "scan_results_UP.txt" and then displayed to the user using the "cat" command.