Add files via upload
This is a bash script that uses the nmap tool to perform a more thorough scan of a target network based on the IP address or range provided by the user. The script prompts the user for input of the target IP or range using the "read" command and stores it in the "target" variable.

The nmap tool is then used to scan the target network with the "-sC" option, which enables running default scripts on the ports found open, the "-sV" option, which tries to detect the service and application versions running on the open ports, and the "-p-" option, which scans all ports instead of just the commonly used ones. The results of the scan are saved to a file named "scan_results_.txt" using the "-oN" option.

Finally, the script displays the scan results to the user using the "cat" command.