Add files via upload
This is a bash script that uses the "nc" (netcat) tool to scan a target host for open ports based on a port range specified by the user. The script prompts the user for input of the target host and port range using the "read" command and stores them in the "host" and "ports" variables, respectively.

The "nc" tool is then used with the "-z" option, which specifies that no data will be sent, and the "-v" option, which enables verbose output, to scan the target host for open ports in the specified range. The results of the scan are displayed to the user on the command line.