Add files via upload
This is a bash script that performs a series of reconnaissance actions on a target IP address provided by the user.

The script starts by prompting the user to input the target IP address using the "read" command and storing it in the "target" variable.

The "whois" command is then used to perform a WHOIS lookup on the target IP address, with the results being saved to a file named "automated_whois.txt" using the ">" redirection operator.

The "host" command is then used to perform a host lookup on the target IP address, with the results being saved to a file named "automated_host.txt" using the ">" redirection operator.

The "nslookup" command is then used to perform an nslookup on the target IP address, with the results being saved to a file named "automated_nslookup.txt" using the ">" redirection operator.

The "traceroute" command is then used to perform a traceroute on the target IP address, with the results being saved to a file named "automated_tracerout.txt" using the ">" redirection operator.

The "nmap" command is then used to perform an nmap scan on the target IP address with service and version detection enabled, and all ports scanned using the "-sC", "-sV", and "-p-" options, respectively. The results of the scan are appended to a file named "automated_nmap.txt" using the ">>" redirection operator.

Finally, the "gobuster" command is used to perform a directory bruteforce attack on the target IP address using a medium-sized wordlist and the results are saved to a file named "automated_gobuster.txt" using the ">>" redirection operator.

Once all the reconnaissance actions are completed, the script outputs the message "Recon run complete!" to the console.