Add files via upload
This script is a Bash script designed to perform a series of reconnaissance actions on a given target domain name. The script prompts the user to input the target domain name, then runs several commands to gather information about the target domain.

The script starts by using the host command to perform a DNS lookup on the target domain and save the output to a file called o.host.txt. It then uses the nslookup command to perform another DNS lookup and save the output to a file called o.nslookup.txt. The script also performs a traceroute using the traceroute command, saves the output to a file called o.traceroute.txt, and uses the dnsrecon command to gather DNS information and save the output to a file called o.dnsrecon.txt.

Additionally, the script uses the wafw00f command to detect the presence of a web application firewall, whois to obtain WHOIS information about the domain, whatweb to gather information about web technologies used on the target domain, and amass to gather subdomain information.

The script then uses the dig command to obtain the IP address of the target domain and performs an nmap scan with service and version detection on the IP address, saving the output to a file called o.nmap.txt.

The script then merges all output files into a single file called automated_recon_$target.txt and displays the contents of the merged file. Finally, it deletes all output files using the rm command.