Add files via upload
The script will perform an Nmap scan with version detection on all ports of the given IP address, save the output to a file called nmap-scan.txt, extract the software names from the output, and then use searchsploit to check for known exploits for each software name. The results of each searchsploit command will be printed to the console.